requirements.txt
streamlit=1.11.0
torch
torchvision
numpy
opencv-python-headless
PyYAML
pillow
timm
tqdm